checkbox and confirm
o proceed, please check the following boxes:
I declare that this property is legitimate.
I agree to the general terms and conditions. with a ref to the page.
You can only proceed if both checkboxes are selected.

Additionally, if you select both check-in and check-out options, you will be able to continue.