🚧✨ Onboarding navigation (WIP)
Added changes from old acceptance branch. Changes from old branch: Created new host onboarding screen with a header component. Added a check if the user wants to leave this screen. Added navigation between pages and keep track of when a user visited a page and also tracks if the input is correct.