BookingOverview Refactor 3/3 - Return user to a website which checks the paymentintent
There are some minor adjustments needed here and there, but the user is getting redirected to a page which checks their payment. The POST function also saves the payment data in the table, which was surprisingly enough forgot to be implemented. Small adjustments to a PATCH request were made as well, but not final.